+++ Packages/systems.speckle.speckle-unity/package.json
@@ -1,9 +1,9 @@
 {
   "name": "systems.speckle.speckle-unity",
-  "version": "2.14.2",
+  "version": "2.15.0",
   "displayName": "Speckle Unity Connector",
   "description": "AEC Interoperability for Unity through Speckle",
-  "unity": "2018.4",
+  "unity": "2020.1",
   "documentationUrl": "https://speckle.guide/user/unity.html",
   "changelogUrl": "https://speckle.systems/blog/",
   "license": "Apache-2.0",
