+++ README.md
@@ -5,12 +5,14 @@
 [![Twitter Follow](https://img.shields.io/twitter/follow/SpeckleSystems?style=social)](https://twitter.com/SpeckleSystems) [![Community forum users](https://img.shields.io/discourse/users?server=https%3A%2F%2Fdiscourse.speckle.works&style=flat-square&logo=discourse&logoColor=white)](https://discourse.speckle.works) [![website](https://img.shields.io/badge/https://-speckle.systems-royalblue?style=flat-square)](https://speckle.systems) [![docs](https://img.shields.io/badge/docs-speckle.guide-orange?style=flat-square&logo=read-the-docs&logoColor=white)](https://speckle.guide/user/unity.html)
 
 
-
 ## Introduction
 
-This repo holds Speckle's Unity Connector + a sample project (Speckle playground). This connector is currently in an Alpha stage.
 
-This connector is meant to be used by developers, it doesn't have an elaborated UI but it offers convenience methods to send and receive data. The connector uses our [Speckle .NET SDK](https://github.com/specklesystems/speckle-sharp).
+This repo holds Speckle's Unity Connector package + a sample project (Speckle playground).
+
+The package offers several Unity Components to send and receive data from Speckle, and allows developers to easily develop their own components and features.
+It has a simple UI, and is missing some of the comforts present in other connectors.
+The connector uses our [Speckle .NET SDK](https://github.com/specklesystems/speckle-sharp).
 
 ![unity](https://user-images.githubusercontent.com/2679513/108543628-3a83ff00-72dd-11eb-8792-3d43ce54e6af.gif)
 
@@ -20,26 +22,28 @@ If you are enjoying using Speckle, don't forget to ⭐ our [GitHub repositories]
 and [join our community forum](https://speckle.community/) where you can post any questions, suggestions, and discuss exciting projects!
 
 ## Notice
-We support Unity 2020 and 2021 (newer versions likely work, but aren't currently part of our test pipeline).
-
+We officially support Unity 2021.3 or newer.
 
 Features:
  - Receive Speckle Objects at Editor or Runtime
- - Send Speckle Objects at Runtime (editor support in the works!)
+ - Send Speckle Objects at Editor or Runtime
  - Material override/substitution
  - Automatic receiving changes
  
-Currently tested on Windows and MacOS. Experimental support for Android [in the works](https://github.com/specklesystems/speckle-unity/issues/68).
+Currently tested on Windows, Linux, and MacOS.
 
-## Sample project
-This repo holds a simple sample project (Speckle Playground). Simply [download this repo](https://github.com/specklesystems/speckle-unity/archive/refs/heads/main.zip)
-or clone with git, and open in Unity 2020.3.
+Android will work [with some signficant limitations](https://github.com/specklesystems/speckle-unity/issues/68), and other platforms likly work with similar limitations.
+
+## Sample Project
+This repo holds a simple sample project (Speckle Playground), containing an example GUI (UnityUI) for fetching stream/branch data, and sending/receiving geometry to/from Speckle.
+
+Simply [download this repo](https://github.com/specklesystems/speckle-unity/archive/refs/heads/main.zip)
+or clone with git, and open in Unity 2021.3 or newer.
 ```
 git clone https://github.com/specklesystems/speckle-unity.git
 ```
-The sample project contains an example GUI (UnityUI) for fetching stream/branch data, and receiving/sending geometry to Speckle.
 
-## Installation
+## Installation (Package)
 
 To install the connector into your own Unity project (rather than using the sample project), open the Package Manager (`Windows -> Package Manager`)
 and select **Add Package from git URL**. (requires [git](https://git-scm.com/downloads) installed)
@@ -59,18 +63,18 @@ We encourage everyone interested to hack / contribute / debug / give feedback to
 
 ### Requirements
 
-- Unity 2020.3+
+- Unity 2021 or greater
 - Have created an account on [speckle.xyz](https://speckle.xyz) (or your own server)
 - Installed [Speckle Manager](https://speckle.guide/user/manager.html) (recommended, otherwise you'll need to implement your own authentication system in Unity)
 
 ### Dependencies
 
-All dependencies to Speckle Core have been included; compiled in the Asset folder until we figure out how to best reference Core.
+All dependencies to Speckle Core have been included; compiled in `systems.speckle.speckle-unity` package.
 
 
 ## Contributing
 
-Please make sure you read the [contribution guidelines](.github/CONTRIBUTING.md) for an overview of the best practices we try to follow.
+Please make sure you read the [contribution guidelines](https://github.com/specklesystems/speckle-sharp/blob/main/.github/CONTRIBUTING.md) for an overview of the best practices we try to follow.
 
 
 ## License
